refactor(prime-affine): isolate local arithmetic contracts
