+++ src/jacobian/math/prime_affine_forms/_operations.py
@@ -2,23 +2,18 @@
 
 from __future__ import annotations
 
-from fractions import Fraction
 from math import prod
 
-from jacobian._exact import CanonicalRational
 from jacobian.canonical import format_canonical_integer, parse_canonical_integer
 from jacobian.math.prime_affine_forms._kernel import (
     interval_match_summary,
     interval_matches,
-    local_bad_residues,
-    local_factor_from_bad_count,
-    primes_through,
     translated_tuple,
     wheel_modulus,
     wheel_rows,
 )
+from jacobian.math.prime_affine_forms._local_factors import local_summary
 from jacobian.math.prime_affine_forms._models import (
-    FinitePrimeTupleFactorProduct,
     PrimeAffineIntervalCountRequest,
     PrimeAffineIntervalEnumerateRequest,
     PrimeAffineTranslationRequest,
@@ -26,17 +21,10 @@ from jacobian.math.prime_affine_forms._models import (
     PrimePatternIntervalCountResult,
     PrimePatternIntervalEnumerateResult,
     PrimePatternMatch,
-    PrimeTupleAdmissibilityRequest,
-    PrimeTupleAdmissibilityResult,
-    PrimeTupleBadResidueRow,
+)
+from jacobian.math.prime_affine_forms._residue_wheel import (
     PrimeTupleIntervalResidueProfileRequest,
     PrimeTupleIntervalResidueProfileResult,
-    PrimeTupleLocalFactorRequest,
-    PrimeTupleLocalFactorResult,
-    PrimeTupleLocalFactorRow,
-    PrimeTupleLocalFactorsRequest,
-    PrimeTupleLocalSummary,
-    PrimeTupleResidueRow,
     PrimeTupleResidueWheel,
     PrimeTupleResidueWheelEnumeration,
     PrimeTupleResidueWheelEnumerationRequest,
@@ -45,99 +33,12 @@ from jacobian.math.prime_affine_forms._models import (
     PrimeTupleWheelMembershipResult,
     PrimeTupleWheelResidueRow,
 )
-from jacobian.math.prime_affine_forms.values import PrimeAffineTuple
-
-
-def _summary(source: PrimeAffineTuple, prime: int) -> PrimeTupleLocalSummary:
-    bad = local_bad_residues(source, prime)
-    return PrimeTupleLocalSummary(
-        prime=prime,
-        bad_residues=tuple(
-            PrimeTupleBadResidueRow(residue=residue, form_ids=form_ids)
-            for residue, form_ids in bad
-        ),
-        bad_count=len(bad),
-        valid_count=prime - len(bad),
-    )
-
-
-def compute_local_factor(
-    request: PrimeTupleLocalFactorRequest,
-) -> PrimeTupleLocalFactorResult:
-    bad = dict(local_bad_residues(request.source, request.prime))
-    return PrimeTupleLocalFactorResult(
-        source=request.source,
-        prime=request.prime,
-        residue_rows=tuple(
-            PrimeTupleResidueRow(
-                residue=residue,
-                vanishing_form_ids=bad.get(residue, ()),
-            )
-            for residue in range(request.prime)
-        ),
-        bad_count=len(bad),
-        valid_count=request.prime - len(bad),
-        locally_obstructed=len(bad) == request.prime,
-        factor=CanonicalRational.from_fraction(
-            local_factor_from_bad_count(
-                request.source.form_count, request.prime, len(bad)
-            )
-        ),
-    )
-
-
-def compute_local_factors(
-    request: PrimeTupleLocalFactorsRequest,
-) -> FinitePrimeTupleFactorProduct:
-    product = Fraction(1, 1)
-    rows: list[PrimeTupleLocalFactorRow] = []
-    first_obstruction: int | None = None
-    for prime in request.primes:
-        summary = _summary(request.source, prime)
-        factor = local_factor_from_bad_count(
-            request.source.form_count, prime, summary.bad_count
-        )
-        rows.append(
-            PrimeTupleLocalFactorRow(
-                summary=summary,
-                factor=CanonicalRational.from_fraction(factor),
-            )
-        )
-        product *= factor
-        if first_obstruction is None and summary.valid_count == 0:
-            first_obstruction = prime
-    return FinitePrimeTupleFactorProduct(
-        source=request.source,
-        primes=request.primes,
-        rows=tuple(rows),
-        product=CanonicalRational.from_fraction(product),
-        first_obstructing_prime=first_obstruction,
-    )
-
-
-def compute_local_admissibility(
-    request: PrimeTupleAdmissibilityRequest,
-) -> PrimeTupleAdmissibilityResult:
-    cutoff = request.source.form_count
-    checked_primes = primes_through(cutoff)
-    rows = tuple(_summary(request.source, prime) for prime in checked_primes)
-    obstructing = tuple(row.prime for row in rows if row.valid_count == 0)
-    return PrimeTupleAdmissibilityResult(
-        source=request.source,
-        cutoff=cutoff,
-        checked_primes=checked_primes,
-        local_rows=rows,
-        status="LOCALLY_OBSTRUCTED" if obstructing else "LOCALLY_ADMISSIBLE",
-        least_obstructing_prime=obstructing[0] if obstructing else None,
-        large_prime_lower_bound=cutoff + 1,
-        maximum_large_prime_bad_residues=request.source.form_count,
-    )
 
 
 def compute_residue_wheel(
     request: PrimeTupleResidueWheelRequest,
 ) -> PrimeTupleResidueWheel:
-    local_rows = tuple(_summary(request.source, prime) for prime in request.primes)
+    local_rows = tuple(local_summary(request.source, prime) for prime in request.primes)
     return PrimeTupleResidueWheel(
         source=request.source,
         primes=request.primes,
@@ -275,9 +176,6 @@ __all__ = [
     "compute_interval_count",
     "compute_interval_enumerate",
     "compute_interval_residue_profile",
-    "compute_local_admissibility",
-    "compute_local_factor",
-    "compute_local_factors",
     "compute_residue_wheel",
     "compute_residue_wheel_enumeration",
     "compute_translation",
