+++ tests/math/prime_affine_forms/test_prime_affine_forms.py
@@ -13,37 +13,43 @@ from jacobian.math.prime_affine_forms import (
     PrimeAffineTuple,
     PrimitiveIntegerAffineForm,
 )
-from jacobian.math.prime_affine_forms._models import (
+from jacobian.math.prime_affine_forms._admissibility import (
+    PrimeTupleAdmissibilityRequest,
+    compute_local_admissibility,
+)
+from jacobian.math.prime_affine_forms._local_factors import (
     FinitePrimeTupleFactorProduct,
+    PrimeTupleLocalFactorRequest,
+    PrimeTupleLocalFactorResult,
+    PrimeTupleLocalFactorsRequest,
+    compute_local_factor,
+    compute_local_factors,
+)
+from jacobian.math.prime_affine_forms._models import (
     PrimeAffineIntervalCountRequest,
     PrimeAffineIntervalEnumerateRequest,
     PrimeAffineTranslationRequest,
     PrimeAffineTranslationResult,
     PrimePatternIntervalCountResult,
     PrimePatternIntervalEnumerateResult,
-    PrimeTupleAdmissibilityRequest,
-    PrimeTupleIntervalResidueProfileRequest,
-    PrimeTupleLocalFactorRequest,
-    PrimeTupleLocalFactorResult,
-    PrimeTupleLocalFactorsRequest,
     PrimeTupleLocalSummary,
-    PrimeTupleResidueWheel,
-    PrimeTupleResidueWheelEnumeration,
-    PrimeTupleResidueWheelEnumerationRequest,
-    PrimeTupleResidueWheelRequest,
-    PrimeTupleWheelMembershipRequest,
 )
 from jacobian.math.prime_affine_forms._operations import (
     compute_interval_count,
     compute_interval_enumerate,
     compute_interval_residue_profile,
-    compute_local_admissibility,
-    compute_local_factor,
-    compute_local_factors,
     compute_residue_wheel,
     compute_residue_wheel_enumeration,
     compute_translation,
     compute_wheel_membership,
+)
+from jacobian.math.prime_affine_forms._residue_wheel import (
+    PrimeTupleIntervalResidueProfileRequest,
+    PrimeTupleResidueWheel,
+    PrimeTupleResidueWheelEnumeration,
+    PrimeTupleResidueWheelEnumerationRequest,
+    PrimeTupleResidueWheelRequest,
+    PrimeTupleWheelMembershipRequest,
 )
 from jacobian.math.prime_affine_forms._tools import TOOLS
 from jacobian.math.prime_affine_forms.values import MAX_AFFINE_AGGREGATE_DIGITS
